chore(rename): renamed file with lesson detail dialog from bottom sheet to general detail screen
