+++ lib/ui/timetable/pages/timetable_week/lessontile_bottomsheet/timetable_lessontile_details_screen.dart
